+++ frontend/src/api/files.ts
@@ -99,7 +99,7 @@ export function download(format: any, ...files: string[]) {
     url += `algo=${format}&`;
   }
 
-  window.open(url);
+  window.open(url, "_blank", "noopener");
 }
 
 export async function post(