+++ frontend/src/views/files/Preview.vue
@@ -131,7 +131,12 @@
             {{ $t("files.noPreview") }}
           </div>
           <div>
-            <a target="_blank" :href="downloadUrl" class="button button--flat">
+            <a
+              target="_blank"
+              rel="noopener noreferrer"
+              :href="downloadUrl"
+              class="button button--flat"
+            >
               <div>
                 <i class="material-icons">file_download</i
                 >{{ $t("buttons.download") }}
@@ -139,6 +144,7 @@
             </a>
             <a
               target="_blank"
+              rel="noopener noreferrer"
               :href="previewUrl"
               class="button button--flat"
               v-if="!fileStore.req?.isDir"
@@ -494,8 +500,8 @@ const close = () => {
   router.push({ path: uri });
 };
 
-const download = () => window.open(downloadUrl.value);
-const openDirect = () => window.open(directUrl.value);
+const download = () => window.open(downloadUrl.value, "_blank", "noopener");
+const openDirect = () => window.open(directUrl.value, "_blank", "noopener");
 
 const editAsText = () => {
   router.push({ path: route.path, query: { edit: "true" } });